+++ src/Boundary.cpp
@@ -87,13 +87,10 @@ void Boundary::set_bounds_to_terrain_pc(Point_set_3& pointCloud, const Polygon_2
     //-- Remove points out of the boundary region
     Boundary::set_bounds_to_buildings_pc(pointCloud, pcBndPoly);
 
-    //-- Add outer points to match the domain size to prescribed one
-    SearchTree searchTree(pointCloud.points().begin(),pointCloud.points().end());
-
     std::vector<double> bndHeights;
     geomutils::interpolate_poly_from_pc(bndPoly, bndHeights, pointCloud);
     _outerBndHeight = geomutils::avg(bndHeights); // Height for buffer (for now) - average of outer pts
-    Config::get().logSummary << "Domain edge height: " << _outerBndHeight << std::endl;
+    Config::get().logSummary << "Domain edge elevation: " << _outerBndHeight << std::endl;
 
     if (Config::get().domainBuffer > -global::largnum + global::smallnum) {
         /*