+++ src/PointCloud.h
@@ -34,20 +34,17 @@
 #include "lasreader.hpp"
 #include "CSF/src/CSF.h"
 
-typedef std::shared_ptr<SearchTree> SearchTreePtr;
-
 class PointCloud {
 public:
     PointCloud();
     ~PointCloud();
 
     void random_thin_pts();
-    void create_flat_terrain(const PolyFeatures& lsFeatures);
+    void create_flat_terrain(const PolyFeaturesPtr& lsFeatures);
     void set_flat_terrain();
     void smooth_terrain();
-    void flatten_polygon_pts(const PolyFeatures& lsFeatures, std::list<Polygon_3>& constrainedPolys);
-    void buffer_flat_edges(const PolyFeatures& avgFeatures, std::list<Polygon_3>& constrainedPolys);
-    SearchTreePtr make_search_tree_buildings();
+    void flatten_polygon_pts(const PolyFeaturesPtr& lsFeatures, std::list<Polygon_3>& constrainedPolys);
+    void buffer_flat_edges(const PolyFeaturesPtr& avgFeatures, std::list<Polygon_3>& constrainedPolys);
     void read_point_clouds();
 
     Point_set_3& get_terrain();